--- libcanard/tests/all.lua
@@ -1,4 +1,0 @@
-require 'tests/test_crc16'
-require 'tests/test_ds015'
-require 'tests/test_libcanard'
-require 'tests/test_ds015'
